Jitterfit app conversion, test, and test data
