+++ isis/src/clipper/apps/jitterfit/jitterfit.h
@@ -1,0 +1,19 @@
+#ifndef jitterfit_h
+#define jitterfit_h
+
+/** This is free and unencumbered software released into the public domain.
+
+The authors of ISIS do not claim copyright on the contents of this file.
+For more details about the LICENSE terms and the AUTHORS, you will
+find files of those names at the top level of this repository. **/
+
+/* SPDX-License-Identifier: CC0-1.0 */
+
+#include "Cube.h"
+#include "UserInterface.h"
+
+namespace Isis{
+  extern void jitterfit(UserInterface &ui);
+}
+
+#endif
